Adding all the new games!
4 games have been added.

The first of these is Forged. It is the first python game I made, and uses rudimentary systems, but is a fun prototype. Because I didn't know what was going on, it is poorly documented.

The second game I added was Platformer Speedrunner. It requires Pygame, but the goal is for the blue block to reach the green without touching the red. It is somewhat polished, but could always use some tinkering.

The gambler's game is one of my favorites. It was incredibly addictive to some of the people I had test it. it is also very polished and partially documented.

The last game is Zombies. I will probably edit this in the future, as it is quickly built, and with little polish. It is still fun, and has some potential. It also requires Pygame.

All together, this is a fun bunch of games!